Rename serializers to serializer to be more consistent
